requirements.txt
certifi==2020.12.5
cycler==0.10.0
glob2==0.7
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
olefile==0.46
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
PySide2==5.15.2
python-dateutil==2.8.1
shiboken2==5.15.2
six==1.15.0
wincertstore==0.2
